Moves build & rebuild to after user creation
If the users do not already exist then installers who already have a configuration containing

```nix
services.nix-daemon.enable = false;
```

will not be able to complete an initial bootstrap on a clean machine. This is because the bootstrap script currently only creates users after an initial rebuild.

The users must exist before the rebuild is completed or the whole thing fails.