examples: drop `ofborg` example
We decided to drop this example as the package is not available in
Nixpkgs and we won't be able to import it easily and keep this example
evaluating as a useful smoke test. The code in this example is already
documented under `services.ofborg.*` so any interested users can still
find out how to set up `ofborg`.