checks: remove user‐dependent checks
I’m not sure how important the channel check is given the
`nixPath` check, and support for managing single‐user installs
isn’t long for this world. These are mostly of limited utility in a
post‐user‐activation world, and the adaptation they’d require
just to keep them around a bit longer for compatibility doesn’t
seem worth it.