Add support for extraModules input
NixOS offers the possibility to pass a list of extraModules to
lib.nixosSystem which later get merged with the modules list and
baseModules.
This offers the possibility to read additional modules from environment
files or a custom module-list.nix.

NixOS implementation is
[here](https://github.com/NixOS/nixpkgs/blob/8601973cec591ab18357eb7c26c52769d990d1ed/nixos/lib/eval-config.nix#L33).

Dor nix-darwin an implementation was started in #592.
I restarted the implementation and removed the usage of
`NIXOS_EXTRA_MODULE_PATH` env variable because NixOS deprecated this
one.