Add NSStatusItemSpacing and NSStatusItemSelectionPadding
Add `NSStatusItemSpacing` and `NSStatusItemSelectionPadding` to `NSGlobalDomain` options.

These options control the spacing between and padding inside status icons in the menu bar. With these options, it's possible to squeeze more items on the menu bar, something that's especially useful on machines with a notch.

This is a copy of #872, which was closed without being merged.