Add exp-features option to run commands
As of early 2025, after a fresh nix install, `flakes` and `nix-command` are active. One can manually activate them in `/etc/nix/nix.conf` but that will cause a conflict during the `nix-darwin` install, which might be confusing to beginners. 

Assuming most people will follow this tutorial after coming from a fresh nix install, having the options inlined in the example command is probably the best approach.